# Faster report result index pages