--- a/app/helpers/application_helper/button/report_download_choice.rb
+++ b/app/helpers/application_helper/button/report_download_choice.rb
@@ -1,5 +1,5 @@
 class ApplicationHelper::Button::ReportDownloadChoice < ApplicationHelper::Button::Basic
   def disabled?
-    MiqReportResult.find(@report_result_id).try(:miq_report_result_details).try(:length).to_i == 0
+    !MiqReportResult.find(@report_result_id).try(:miq_report_result_details).try(:exists?)
   end
 end