@NickLaMuro 2018-06-18
@jrafanie Oh, these ones are on purpose.  We need the stuff merged in on `ManageIQ/manageiq` for these to function.  Hence the test failures.

There might be something up with the `javascript` tests, but my guess is it is either related to the missing methods on master, or something that isn't my fault (my story and I am sticking to it 😤 )