@NickLaMuro 2018-06-19
Current spec failures are do to issue on `ManageIQ/manageiq:master`.  Fix here:

https://github.com/ManageIQ/manageiq/pull/17605