@NickLaMuro 2018-06-19
@jrafanie Finally.... 😩 

Think this is good to go now.