@simaishi 2018-06-22
@NickLaMuro Can you create a separate PR for Fine branch as there are good amount of conflicts on backport?

```
$ git status
On branch fine
Your branch is up-to-date with 'origin/fine'.
You are currently cherry-picking commit 49a3cb6.
  (fix conflicts and run "git cherry-pick --continue")
  (use "git cherry-pick --abort" to cancel the cherry-pick operation)

Changes to be committed:

	modified:   app/controllers/chargeback_controller.rb
	modified:   app/helpers/application_helper/button/report_download_choice.rb
	modified:   app/helpers/application_helper/button/report_only.rb
	modified:   spec/controllers/chargeback_controller_spec.rb

Unmerged paths:
  (use "git add/rm <file>..." as appropriate to mark resolution)

	deleted by us:   app/controllers/mixins/saved_report_paging.rb
	both modified:   app/controllers/report_controller/saved_reports.rb
	both modified:   spec/controllers/miq_report_controller/trees_spec.rb
	both modified:   spec/controllers/report_controller_spec.rb
```