@simaishi 2018-06-22
@NickLaMuro For Gaprindashvili, there is only one conflict, but I don't know how to resolve...

spec/controllers/miq_report_controller/trees_spec.rb 

```diff
      it 'renders show' do
<<<<<<< HEAD
        user = FactoryGirl.create(:user_with_group)
        login_as user
=======
        session[:settings] = {:perpage => {:reports => 20}}
>>>>>>> 49a3cb6... Merge pull request #4143 from NickLaMuro/faster_report_result_display
        controller.instance_variable_set(:@html, "<h1>Test</h1>")
```

So can you either tell me how it should be resolved 😄  or create a PR?