@simaishi 2018-06-26
Backported to Gaprindashvili via https://github.com/ManageIQ/manageiq-ui-classic/pull/4143