@simaishi 2018-06-26
Backported to Fine via https://github.com/ManageIQ/manageiq-ui-classic/pull/4208